requirements.txt
numpy
pandas
cython<=0.30
pyzmq
cython-dev-tools
pytest
